Move remaining preview code
