+++ src/app/components/previews/BiomeSourcePreview.tsx
@@ -4,13 +4,13 @@ import type { mat3 } from 'gl-matrix'
 import { useCallback, useRef, useState } from 'preact/hooks'
 import { getProjectData, useLocale, useProject, useStore } from '../../contexts/index.js'
 import { useAsync } from '../../hooks/index.js'
-import { DEEPSLATE } from '../../previews/Deepslate.js'
 import { Store } from '../../Store.js'
 import { iterateWorld2D, randomSeed, stringToColor } from '../../Utils.js'
 import { Btn, BtnMenu, NumberInput } from '../index.js'
 import type { ColormapType } from './Colormap.js'
 import { getColormap } from './Colormap.js'
 import { ColormapSelector } from './ColormapSelector.jsx'
+import { DEEPSLATE } from './Deepslate.js'
 import type { PreviewProps } from './index.js'
 import { InteractiveCanvas2D } from './InteractiveCanvas2D.jsx'
