+++ src/app/components/previews/Deepslate.ts
@@ -1,7 +1,7 @@
 import * as deepslate19 from 'deepslate/worldgen'
-import type { VersionId } from '../services/index.js'
-import { checkVersion, fetchAllPresets, fetchPreset } from '../services/index.js'
-import { clamp, computeIfAbsent, computeIfAbsentAsync, deepClone, deepEqual, isObject, square } from '../Utils.js'
+import type { VersionId } from '../../services/index.js'
+import { checkVersion, fetchAllPresets, fetchPreset } from '../../services/index.js'
+import { clamp, computeIfAbsent, computeIfAbsentAsync, deepClone, deepEqual, isObject, square } from '../../Utils.js'
 
 export type ProjectData = Record<string, Record<string, unknown>>
