+++ src/app/components/previews/NoiseSettingsPreview.tsx
@@ -5,7 +5,6 @@ import { vec2 } from 'gl-matrix'
 import { useCallback, useMemo, useRef, useState } from 'preact/hooks'
 import { getProjectData, useLocale, useProject } from '../../contexts/index.js'
 import { useAsync } from '../../hooks/index.js'
-import { DEEPSLATE } from '../../previews/Deepslate.js'
 import { CachedCollections } from '../../services/index.js'
 import { Store } from '../../Store.js'
 import { iterateWorld2D, randomSeed } from '../../Utils.js'
@@ -13,6 +12,7 @@ import { Btn, BtnInput, BtnMenu } from '../index.js'
 import type { ColormapType } from './Colormap.js'
 import { getColormap } from './Colormap.js'
 import { ColormapSelector } from './ColormapSelector.jsx'
+import { DEEPSLATE } from './Deepslate.js'
 import type { PreviewProps } from './index.js'
 import { InteractiveCanvas2D } from './InteractiveCanvas2D.jsx'
