--- src/app/previews/index.ts
@@ -1,1 +1,0 @@
-export * from './Decorator.js'
